Some cosmetic changes

+++ examples/external_class/library/Printer/source/PrinterImpl.h
@@ -43,9 +43,9 @@ namespace PrinterNS
             std::string result = "";
             int indent = 0;
             if (mAlign == CompC::Right)
-                indent = 80 - text.length();
+                indent = 80 - static_cast<int>(text.length());
             else if (mAlign == CompC::Center)
-                indent = (80 - text.length()) / 2;
+                indent = (80 - static_cast<int>(text.length())) / 2;
             for (int i = 0; i < indent; ++i)
                 result += " ";
             result += text;
